Refactor models to use decimal types and Wallet value object
Changed financial model properties from double to decimal for improved precision. Introduced Wallet value object and updated controllers, mappers, services, and repository interfaces to use Wallet instead of string. Renamed and refactored PositionModel to Position and TradeModel to Trade. Added new models for wallet position snapshots and minimum PE for trades. Updated dependencies and added .editorconfig. Added initial unit test project for wallet position service.

+++ CopyTrading/Mappers/CandleMapper.cs
@@ -10,12 +10,12 @@ public static class CandleMapper
     {
         return new Candle()
         {
-            Open = (double)hlKline.OpenPrice,
-            High = (double)hlKline.HighPrice,
-            Low = (double)hlKline.LowPrice,
-            Close = (double)hlKline.ClosePrice,
+            Open = (decimal)hlKline.OpenPrice,
+            High = (decimal)hlKline.HighPrice,
+            Low = (decimal)hlKline.LowPrice,
+            Close = (decimal)hlKline.ClosePrice,
             Symbol = hlKline.Symbol,
-            Volume = (double)hlKline.Volume,
+            Volume = (decimal)hlKline.Volume,
             Time = hlKline.OpenTime,
             Interval = (int)hlKline.Interval / 60,
         };
