+++ CopyTrading/Models/Candle.cs
@@ -3,12 +3,12 @@ namespace CopyTrading.Models;
 
 public class Candle
 {
-    public double Open { get; set; }
-    public double High { get; set; }
-    public double Low { get; set; }
-    public double Close { get; set; }
+    public decimal Open { get; set; }
+    public decimal High { get; set; }
+    public decimal Low { get; set; }
+    public decimal Close { get; set; }
     public string Symbol { get; set; }
-    public double Volume { get; set; }
+    public decimal Volume { get; set; }
     public DateTime Time { get; set; }
     public int Interval { get; set; }
 }
